+++ easyrec-core/src/main/java/org/easyrec/store/dao/core/impl/RecommendedItemDAOMysqlImpl.java
@@ -317,7 +317,7 @@ public class RecommendedItemDAOMysqlImpl
                             DaoUtils.getInteger(rs, DEFAULT_ITEM_TYPE_COLUMN_NAME)),
                     DaoUtils.getDouble(rs, DEFAULT_PREDICTION_VALUE_COLUMN_NAME),
                     DaoUtils.getInteger(rs, DEFAULT_RECOMMENDATION_COLUMN_NAME),
-                    DaoUtils.getInteger(rs, DEFAULT_ITEM_ASSOC_COLUMN_NAME),
+                    DaoUtils.getLong(rs, DEFAULT_ITEM_ASSOC_COLUMN_NAME),
                     DaoUtils.getStringIfPresent(rs, DEFAULT_EXPLANATION_COLUMN_NAME));
             return recommendedItem;
         }
@@ -337,7 +337,7 @@ public class RecommendedItemDAOMysqlImpl
                             DaoUtils.getInteger(rs, DEFAULT_ITEM_TYPE_COLUMN_NAME)),
                     DaoUtils.getDouble(rs, DEFAULT_PREDICTION_VALUE_COLUMN_NAME),
                     DaoUtils.getInteger(rs, DEFAULT_RECOMMENDATION_COLUMN_NAME),
-                    DaoUtils.getInteger(rs, DEFAULT_ITEM_ASSOC_COLUMN_NAME),
+                    DaoUtils.getLong(rs, DEFAULT_ITEM_ASSOC_COLUMN_NAME),
                     DaoUtils.getStringIfPresent(rs, DEFAULT_EXPLANATION_COLUMN_NAME));
             return recommendedItem;
         }
